requirements.txt
certifi==2022.12.7
charset-normalizer==3.0.1
idna==3.4
pyaes==1.6.1
pyasn1==0.4.8
python-dotenv==0.21.1
requests==2.28.2
rsa==4.9
Telethon==1.27.0
urllib3==1.26.14
vk-api==11.9.9
